improvements of spp_land_record and spp_farmer_registry_base modules
